requirements.txt
fastapi
uvicorn
python-docx
PyMuPDF
language-tool-python
python-multipart
